% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{add_economics,RSVProgramme-method}
\alias{add_economics,RSVProgramme-method}
\title{A setMethod for the \code{add_economics} function}
\usage{
\S4method{add_economics}{RSVProgramme}(object, econ_name, econ_raw, risks_raw, risks_vhr_raw)
}
\arguments{
\item{econ_name}{a string for the name gven the economic framework given}

\item{econ_raw}{a dataframe with the economic parameters per age group.}

\item{risks_raw}{a dataframe with the risk of outcomes per age group.}

\item{risks_vhr_raw}{a dataframe with the risk of outcomes per age group for very-high-risk infants}
}
\value{
An RSVProgramme object
}
\description{
A setMethod for the \code{add_economics} function
}
