% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{add_economics}
\alias{add_economics}
\title{Add Economics Data to an Object}
\usage{
add_economics(object, econ_name, econ_raw, risks_raw, risks_vhr_raw)
}
\arguments{
\item{object}{The object to which the economics data will be added.}

\item{econ_name}{A character string representing the name of the economics data.}

\item{econ_raw}{The raw economics data to be added.}

\item{risks_raw}{The raw risks data associated with the economics data.}

\item{risks_vhr_raw}{The raw very high risk data associated with the economics data.}
}
\value{
The object with the added economics data.
}
\description{
This generic function is used to add economics data to a specified object.
}
