% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_economics_methods.R
\name{add_fitted_dist}
\alias{add_fitted_dist}
\title{Add Fitted Distribution to Data Frame}
\usage{
add_fitted_dist(cost_df, dist)
}
\arguments{
\item{cost_df}{A data frame containing the data to which the distribution will be fitted. The data frame should have columns named "mean", "lb_95", and "ub_95".}

\item{dist}{A character string specifying the distribution to fit. The distribution should be one of the distributions supported by the \code{q} function in R (e.g., "norm" for normal distribution, "tri" for triangular distribution).}
}
\value{
A data frame with an additional column named \code{fitt_dist} containing the fitted distribution for each row.
}
\description{
This function fits a specified distribution to the data in a given data frame and adds the fitted distribution as a new column.
}
\examples{
\dontrun{
cost_df <- data.frame(mean = c(10, 20), lb_95 = c(5, 15), ub_95 = c(15, 25))
add_fitted_dist(cost_df, "norm")
}

}
