% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{add_programme,RSVProgramme-method}
\alias{add_programme,RSVProgramme-method}
\title{A setMethod for the \code{add_programme} function}
\usage{
\S4method{add_programme}{RSVProgramme}(object, prog_name, cal, cal_vhr, immune_profile)
}
\arguments{
\item{prog_name}{an integer taking the number of Monte Carlo samples to run over}

\item{cal}{a dataframe with the implementation calendar for the population}

\item{cal_vhr}{dataframe with the implementation calendar for the very-high-risk population}

\item{immune_profile}{a list with information about the immune asumptions if the prophylatic in the programme}
}
\value{
An RSVProgramme object
}
\description{
A setMethod for the \code{add_programme} function
}
