% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{add_programme}
\alias{add_programme}
\title{Add Programme}
\usage{
add_programme(object, prog_name, cal, cal_vhr, immune_profile)
}
\arguments{
\item{object}{The object to which the programme will be added.}

\item{prog_name}{The name of the programme to be added.}

\item{cal}{The calendar associated with the programme.}

\item{cal_vhr}{The calendar with very high resolution associated with the programme.}

\item{immune_profile}{The immune profile associated with the programme.}
}
\value{
The object with the added programme.
}
\description{
This function is a generic method for adding a programme to an object.
}
