% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_programme_methods.R
\name{calculate_daily_uptake}
\alias{calculate_daily_uptake}
\title{Convert daily uptake from calendar to model}
\usage{
calculate_daily_uptake(object, cal_df)
}
\arguments{
\item{object}{am S4 class of the defined RSV programme}

\item{cal_df}{a dataframe of a calendar being evaluated}
}
\description{
Convert mean and ci into samples from a fitted distribution.
}
