% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkout_states}
\alias{checkout_states}
\title{function to conver the incidence matrix from a state run to a data.frame}
\usage{
checkout_states(object)
}
\arguments{
\item{object}{an RSVProgramme object}
}
\value{
none
}
\description{
function to conver the incidence matrix from a state run to a data.frame
}
