% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_interventions}
\alias{compare_interventions}
\title{Compare Interventions}
\usage{
compare_interventions(obj_base, obj_inter)
}
\arguments{
\item{obj_base}{An object containing the outcomes of the base intervention.}

\item{obj_inter}{A list of objects, each containing the outcomes of different interventions to be compared with the base intervention.}
}
\value{
A data frame summarizing the total cases, cases in very high-risk (VHR) groups, and total cases for each age group and outcome, for both the base and other interventions.
}
\description{
This function compares the outcomes of a base intervention with multiple other interventions.
}
