% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_programme_methods.R
\name{convert_efficacies}
\alias{convert_efficacies}
\title{Convert Immune Profile to Efficacies}
\usage{
convert_efficacies(immune_profile)
}
\arguments{
\item{immune_profile}{A list containing immune profile information. It should have the following structure:
\describe{
\item{vhr}{A list with element \code{b} representing the efficacy for very high risk (VHR) individuals.}
\item{mass}{A list with elements \code{b} and \code{product} representing the efficacy and product type for mass immunization. The \code{product} can be "lav", "mab", or "mat".}
}}
}
\value{
A list with the following elements:
\describe{
\item{mab_vhr}{Efficacy for monoclonal antibodies (MAB) in very high risk individuals.}
\item{mab_mass}{Efficacy for monoclonal antibodies (MAB) in mass immunization.}
\item{lav_mass}{Efficacy for live attenuated vaccines (LAV) in mass immunization.}
\item{mat_mass}{Efficacy for maternal antibodies (MAT) in mass immunization.}
}
}
\description{
This function takes an immune profile and converts it into a list of efficacies for different products.
}
\details{
The function initializes the efficacy list with zeros. It then checks if the immune profile contains efficacy values for VHR and mass immunization and updates the efficacy list accordingly.
}
\examples{
immune_profile <- list(
  vhr = list(b = 0.8),
  mass = list(b = 0.6, product = "lav")
)
convert_efficacies(immune_profile)

}
