% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_programme_methods.R
\name{convert_mat_to_trans_cal}
\alias{convert_mat_to_trans_cal}
\title{Convert Matrices to Transmission Calendar}
\usage{
convert_mat_to_trans_cal(
  all_dose,
  all_sero,
  cal_vhr,
  daily_uptake_vhr,
  cal,
  daily_uptake,
  immune_profile
)
}
\arguments{
\item{all_dose}{A matrix representing all doses.}

\item{all_sero}{A matrix representing all sero values.}

\item{cal_vhr}{A calendar matrix for very high risk (VHR) individuals.}

\item{daily_uptake_vhr}{A matrix representing daily uptake for VHR individuals.}

\item{cal}{A calendar matrix for the general population.}

\item{daily_uptake}{A matrix representing daily uptake for the general population.}

\item{immune_profile}{A list containing immune profiles for VHR and general population.}
}
\value{
A list containing updated \code{dose} and \code{sero} matrices.
}
\description{
This function converts dose and sero matrices into a transmission calendar based on the provided immune profile.
}
\details{
The function first creates a VHR calendar if the immune profile for VHR is not "none". It then calculates the sero and dose models using the \code{get_sero_dose_cals} helper function. The results are stored in the \code{all_sero} and \code{all_dose} matrices.

The \code{get_sero_dose_cals} function calculates the dose and sero models based on the provided calendar, daily uptake, and immune profile. It considers different sero delays and coverage values.
}
