% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_programme_methods.R
\name{get_coverage}
\alias{get_coverage}
\title{Get Coverage}
\usage{
get_coverage(immune_profile_mass, cal)
}
\arguments{
\item{immune_profile_mass}{A data frame containing the immune profile mass information.}

\item{cal}{A data frame containing the calibration data.}
}
\value{
The coverage value. If the product in the immune profile mass is "mat", it returns the mean of the positive coverage values from the calibration data. Otherwise, it returns 0.
}
\description{
This function calculates the coverage based on the immune profile mass and calibration data.
}
\examples{
immune_profile_mass <- data.frame(product = "mat")
cal <- data.frame(cov = c(0.1, 0.2, 0, -0.1))
get_coverage(immune_profile_mass, cal)
}
