% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{make_rsv_programme}
\alias{make_rsv_programme}
\title{A constructor to make an RSVProgramme object}
\usage{
make_rsv_programme(S = 10)
}
\arguments{
\item{S}{an integer taking the number of Monte Carlo samples to run over}
}
\value{
A RSVProgramme object
}
\description{
A constructor to make an RSVProgramme object
}
