% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{make_RunInterventions}
\alias{make_RunInterventions}
\title{Initialise the RunInterventions class}
\usage{
make_RunInterventions(RunInterventions, uk_data_sum, run_burn, run_full)
}
\arguments{
\item{RunInterventions}{An empty RunInterventions class}

\item{uk_data_sum}{an .rda file with all the information needed to fit the UK model}

\item{run_burn}{Burn-in for the model (in days)}

\item{run_full}{Number of days to run the model for post burn-in}
}
\value{
A classRunInterventions which is fully parameterised
}
\description{
Initialise the RunInterventions class
}
