% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_economics_methods.R
\name{output_samples}
\alias{output_samples}
\title{Convert mean and ci into samples from a fitted distribution}
\usage{
output_samples(econ_raw, risks_raw, file.name)
}
\arguments{
\item{econ_raw}{a dataframe with the economic parameters per age group.}

\item{risks_raw}{a dataframe with the risk of outcomes per age group.}
}
\description{
Convert mean and ci into samples from a fitted distribution.
}
