% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_programme_methods.R
\name{plot_calendar}
\alias{plot_calendar}
\title{function to plot the dosing calendar of the RSVprogramme]}
\usage{
plot_calendar(object)
}
\arguments{
\item{object}{an RSVProgramme object}
}
\value{
none
}
\description{
function to plot the dosing calendar of the RSVprogramme]
}
