% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_methods.R
\name{plot_outcomes}
\alias{plot_outcomes}
\title{Plot Outcomes}
\usage{
plot_outcomes(object)
}
\arguments{
\item{object}{An S4 object that contains the program name and outcomes data.}
}
\value{
A ggplot object representing the boxplot of costs and QALY outcomes.
}
\description{
This function loads the run outputs from a specified file, extracts the outcomes data, and creates a boxplot of costs and QALY (Quality-Adjusted Life Year) outcomes. The plot is saved as a PNG file.
}
\details{
The function performs the following steps:
\itemize{
\item Loads the run outputs from a file located in the "outputs" directory.
\item Extracts the outcomes data from the loaded object.
\item Binds the costs and QALY data into a single data frame.
\item Creates a boxplot of the costs and QALY data using ggplot2.
\item Saves the plot as a PNG file in the "figs" subdirectory of the "outputs" directory.
}
}
