% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\docType{class}
\name{RSVProgramme-class}
\alias{RSVProgramme-class}
\title{RSVProgramme Class}
\description{
A class to represent an RSV (Respiratory Syncytial Virus) Programme.
}
\section{Slots}{

\describe{
\item{\code{econ_name}}{A character string representing the economic name.}

\item{\code{prog_name}}{A character string representing the programme name.}

\item{\code{model}}{An object of class refClass representing the model.}

\item{\code{uk_data}}{A list containing UK data.}

\item{\code{model_par}}{A list of model parameters.}

\item{\code{econ_par}}{A list of economic parameters.}

\item{\code{model_calendar}}{A list representing the model calendar.}

\item{\code{dose_calendar}}{A list representing the dose calendar.}

\item{\code{immune_profile}}{A list representing the immune profile.}

\item{\code{econ_df}}{A data.frame containing economic data.}

\item{\code{risks_df}}{A data.frame containing risk data.}

\item{\code{risks_vhr_df}}{A data.frame containing very high risk data.}

\item{\code{outcomes_vec}}{A vector of outcomes.}

\item{\code{post}}{A data.frame containing posterior data.}

\item{\code{seeds}}{A vector of seeds.}

\item{\code{S}}{A numeric value.}

\item{\code{raw_inci}}{A list of raw incidence data.}

\item{\code{wane_func_string}}{A character string representing the waning function.}

\item{\code{wane_func_string_vhr}}{A character string representing the waning function for very high risk.}

\item{\code{outcomes}}{A list of outcomes.}

\item{\code{samples_outcomes}}{A list of sample outcomes.}

\item{\code{samples_outcomes_VHR}}{A list of sample outcomes for very high risk.}

\item{\code{efficacies}}{A vector of efficacies.}

\item{\code{cov_mat}}{A numeric matrix representing the covariance matrix.}

\item{\code{full_output}}{A logical value indicating whether to produce full output.}
}}

