% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{run,RSVProgramme-method}
\alias{run,RSVProgramme-method}
\title{A setMethod for the \code{run} function}
\usage{
\S4method{run}{RSVProgramme}(object, direct = FALSE, filename = NULL)
}
\arguments{
\item{object}{RSVProgramme object}
}
\value{
An RSVProgramme object
}
\description{
A setMethod for the \code{run} function
}
