% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{run}
\alias{run}
\title{Run Method}
\usage{
run(object, direct = FALSE, filename = NULL, yr_num = 2)
}
\arguments{
\item{object}{The object on which the operation is performed.}

\item{direct}{A logical value indicating whether to run the operation directly.}

\item{filename}{An optional character string specifying the filename.}

\item{yr_num}{An integer specifying the number of years, default is 2.}
}
\value{
The result of the operation.
}
\description{
Generic function to execute a specific operation.
}
