% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{run_state,RSVProgramme-method}
\alias{run_state,RSVProgramme-method}
\title{A setMethod for the \code{run_state} function}
\usage{
\S4method{run_state}{RSVProgramme}(object)
}
\arguments{
\item{object}{RSVProgramme object}
}
\value{
a matrix with all the values of the state variables in the dynamic transmission model
}
\description{
A setMethod for the \code{run_state} function
}
