% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{run_state}
\alias{run_state}
\title{Run State Generic Function}
\usage{
run_state(object, S_i = 1, direct = FALSE)
}
\arguments{
\item{object}{An object on which the state is to be run.}

\item{S_i}{An integer representing the state index. Default is 1.}

\item{direct}{A logical value indicating whether to run the state directly. Default is FALSE.}
}
\value{
The result of the state run, depending on the method implementation.
}
\description{
This is a generic function for running a state.
}
