% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarise_outcomes}
\alias{summarise_outcomes}
\title{Summarise Outcomes}
\usage{
summarise_outcomes(x)
}
\arguments{
\item{x}{An object containing the outcomes data.}
}
\value{
A data frame with the total number of cases grouped by \code{s}, \code{outcome}, and \code{age_group}.
}
\description{
This function takes an object \code{x} and summarises the outcomes by grouping
by \code{s}, \code{outcome}, and \code{age_group}, then calculates the total number of cases.
}
