% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/programmeClass.R
\name{add_programme,RSVProgramme-method}
\alias{add_programme,RSVProgramme-method}
\title{Add Programme to RSVProgramme Object}
\usage{
\S4method{add_programme}{RSVProgramme}(object, prog_name, cal, cal_vhr, immune_profile)
}
\arguments{
\item{object}{An object of class \code{RSVProgramme}.}

\item{prog_name}{A character string representing the name of the programme.}

\item{cal}{A calendar object used for the programme.}

\item{cal_vhr}{A calendar object for very high-risk (VHR) groups.}

\item{immune_profile}{A list containing the immune profile information.}
}
\value{
The updated \code{RSVProgramme} object with the new programme added.
}
\description{
This method adds a programme to an RSVProgramme object, setting various attributes and creating necessary directories.
}
\details{
This function sets the programme name, checks and creates necessary directories, initializes dose and sero matrices, calculates daily uptake, converts matrices to transmission calendars, and sets various attributes in the \code{RSVProgramme} object.
}
