% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkout_incidence}
\alias{checkout_incidence}
\title{function to conver the incidence matrix from a run to a data.frame}
\usage{
checkout_incidence(object)
}
\arguments{
\item{object}{an RSVProgramme object}
}
\value{
none
}
\description{
function to conver the incidence matrix from a run to a data.frame
}
