% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkout_states}
\alias{checkout_states}
\title{Checkout States}
\usage{
checkout_states(object)
}
\arguments{
\item{object}{An object containing the states to be processed.}
}
\value{
A data table with columns for time, incidence, age group, risk group, social group, model type, and state name.
}
\description{
This function processes the states of a given object and returns a data table with detailed information about each state.
}
\details{
The function generates a list of state names and categorizes them into risk groups, social groups, and model types. It then runs the state function on the input object and processes the resulting states into a data frame. The data frame is then transformed into a long format and additional columns are added to provide detailed information about each state.
}
